fix(name) rename check_link to ensure_valid_link.
related #137
