+++ backend/server/calendars_bookings_get.py
@@ -4,7 +4,7 @@ from tools.build_response import build_response
 from tools.for_db.work_with_slots import get_slots_by_admin_id_and_booking_id, BookingNotFound
 from tools.for_db.work_with_booking_info import get_booking_info
 from tools.for_db.work_with_links import get_link
-from server.validation.check_link import check_link, LinkHasExpired, LinkNotFound
+from server.validation.ensure_valid_link import ensure_valid_link, LinkHasExpired, LinkNotFound
 
 calendars_bookings_get = Blueprint('calendars_bookings_get', __name__)
 
@@ -13,7 +13,7 @@ calendars_bookings_get = Blueprint('calendars_bookings_get', __name__)
 def get_calendars_bookings(link_id, booking_id):
     try:
         link = get_link(link_id)
-        check_link(link)
+        ensure_valid_link(link)
         slots = get_slots_by_admin_id_and_booking_id(link.admin_id, booking_id)
         booking_info = get_booking_info(booking_id)
     except BookingNotFound as e:
